Добавляет модули JS
